Organized code and finalized search functionality
