--- Login-/Admin Panel/Pass hash.php
@@ -1,4 +1,0 @@
-<?php
-$pass1 = password_hash("s", PASSWORD_DEFAULT);
-
-?>
