--- Login-/Admin Panel/jobs.sql
@@ -1,72 +1,0 @@
--- phpMyAdmin SQL Dump
--- version 5.2.0
--- https://www.phpmyadmin.net/
---
--- Host: 127.0.0.1
--- Generation Time: Oct 22, 2022 at 08:01 AM
--- Server version: 10.4.24-MariaDB
--- PHP Version: 8.1.6
-
-SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
-START TRANSACTION;
-SET time_zone = "+00:00";
-
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!40101 SET NAMES utf8mb4 */;
-
---
--- Database: `job orbitz`
---
-
--- --------------------------------------------------------
-
---
--- Table structure for table `jobs`
---
-
-CREATE TABLE `jobs` (
-  `sno` int(10) NOT NULL,
-  `organization` varchar(100) NOT NULL,
-  `image` varchar(250) NOT NULL,
-  `description` varchar(255) NOT NULL,
-  `posted date` date NOT NULL,
-  `last date` date NOT NULL,
-  `province` varchar(10) NOT NULL,
-  `requirements` varchar(100) NOT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
-
---
--- Dumping data for table `jobs`
---
-
-INSERT INTO `jobs` (`sno`, `organization`, `image`, `description`, `posted date`, `last date`, `province`, `requirements`) VALUES
-(1, 'GMMMC Sukkur', 'random images', 'random description', '2022-10-18', '2022-10-31', 'Sindh', 'Paisa + parat'),
-(2, 'Nadra', 'Not Available', 'nahi he', '2022-10-12', '2022-10-18', '', 'Bachelcor in computer science');
-
---
--- Indexes for dumped tables
---
-
---
--- Indexes for table `jobs`
---
-ALTER TABLE `jobs`
-  ADD PRIMARY KEY (`sno`);
-
---
--- AUTO_INCREMENT for dumped tables
---
-
---
--- AUTO_INCREMENT for table `jobs`
---
-ALTER TABLE `jobs`
-  MODIFY `sno` int(10) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=3;
-COMMIT;
-
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
