--- Login-/Admin Panel/logout.php
@@ -1,9 +1,0 @@
-<?php
-
-    session_start();
-    session_unset();
-    session_destroy();
-
-    header("location: login.php");
-    exit;
-?>
