--- Login-/Admin Panel/setcookie.php
@@ -1,3 +1,0 @@
-<?php
-$_COOKIE();
-?>
