+++ Login-/dbConn/dbconn.php
@@ -1,8 +1,8 @@
 <?php
-    $host = 'localhost';
+    $host = '127.0.0.1';
     $data = 'job_orbitz';
     $user = 'root';
-    $pass = '';
+    $pass = '123456789';
     $chrs = 'utf8mb4';
     $attr = "mysql:host=$host;dbname=$data;charset=$chrs";
     $opts = 
@@ -11,4 +11,11 @@
         PDO:: ATTR_DEFAULT_FETCH_MODE   => PDO::FETCH_ASSOC,
         PDO:: ATTR_EMULATE_PREPARES      => false,
     ];  
+
+    try {
+        $pdo = new PDO($attr, $user, $pass, $opts);
+    } catch (PDOException $e) {
+        throw new PDOException($e->getMessage(), (int)$e->getCode());
+    }
+    
 ?>
